[philox_rand] Add decomps
[ghstack-poisoned]
